remove processing message
This caused confusion since the create event
from the paypal buttons is fired off when
clicking on the credit card option, this
does not bring up a model like the paypal
option, instead populates a div below. So
it was weird that the processing message would
show when the credit card option was chosen.